TID-19 fix: Added in functions for when post content is not changed

+++ app/model/posts.py
@@ -35,24 +35,39 @@ class Post(UserMixin):
         cursor.close()
 
     def edit(self):
-        cursor = mysql.connection.cursor(dictionary=True)
-        sql = "UPDATE post SET title = %s, caption = %s, ingredients = %s, instructions = %s, tag = %s, subtags = %s WHERE id = %s"
-        sql_urls = "INSERT INTO post_url (post_id, url, type) VALUES (%s, %s, %s)"
+        if self.content:
+            # If there's new content to be added in
+            cursor = mysql.connection.cursor(dictionary=True)
+            sql = "UPDATE post SET title = %s, caption = %s, ingredients = %s, instructions = %s, tag = %s, subtags = %s WHERE id = %s"
+            sql_urls = "INSERT INTO post_url (post_id, url, type) VALUES (%s, %s, %s)"
 
-        # Modify post data in the post table
-        cursor.execute(sql, (self.title, self.caption, self.ingredients, self.instructions, self.tag, self.subtags, self.id))
+            # Modify post data in the post table
+            cursor.execute(sql, (self.title, self.caption, self.ingredients, self.instructions, self.tag, self.subtags, self.id))
 
-        # Delete old content
-        sql_del = "DELETE FROM post_url WHERE post_id = %s"
-        cursor.execute(sql_del, (self.id,))
+            # Delete old content
+            sql_del = "DELETE FROM post_url WHERE post_id = %s"
+            cursor.execute(sql_del, (self.id,))
 
-        # Insert new content into the `post_urls` table
-        for entry in self.content:
-            cursor.execute(sql_urls, (self.id, entry['url'], entry['type'],))
+            # Insert new content into the `post_urls` table
+            for entry in self.content:
+                cursor.execute(sql_urls, (self.id, entry['url'], entry['type'],))
 
-        # Commit changes
-        mysql.connection.commit()
-        cursor.close()
+            # Commit changes
+            mysql.connection.commit()
+            cursor.close()
+        else:
+            # If no new content
+            cursor = mysql.connection.cursor(dictionary=True)
+            sql = "UPDATE post SET title = %s, caption = %s, ingredients = %s, instructions = %s, tag = %s, subtags = %s WHERE id = %s"
+
+            # Modify post data in the post table
+            cursor.execute(sql, (self.title, self.caption, self.ingredients, self.instructions, self.tag, self.subtags, self.id))
+
+            # Commit changes
+            mysql.connection.commit()
+            cursor.close()
+
+
 
     
     def fetch_post(post_id):
